fix: holiday-jp-goを依存から除去
ベンチマークと互換性確認のみに使用しているため、別モジュールに移動
