+++ compatible/go.mod
@@ -1,0 +1,17 @@
+module compatible
+
+go 1.22
+
+replace github.com/gahojin/go-holiday-japanese => ../
+
+require (
+	github.com/gahojin/go-holiday-japanese v0.0.0-00010101000000-000000000000
+	github.com/holiday-jp/holiday_jp-go v0.0.0-20220125203534-53124b4cc19c
+	github.com/stretchr/testify v1.11.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
